trying to add a shape
